pyproject.toml
[project]
name = "foundation-data-factory"
version = "0.0.1"
description = "FoundationDataFactory is a streaming-first data execution system for large-scale multimodal foundation model training."
authors = [{ name = "Victor An", email = "victor.duo.an@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic>=2.9.0",
    "pyyaml>=6.0.0",
    "pyarrow>=18.0.0",
    "datasets>=2.21.0",
    "daft>=0.4.0",
    "ray>=2.0.0",
    "textstat>=0.7.12",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://duoan.github.io/foundation-data-factory/"
Repository = "https://github.com/duoan/foundation-data-factory"
Documentation = "https://duoan.github.io/foundation-data-factory/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.0",
    "tox-uv>=1.29.0",
    "deptry>=0.24.0",
    "ty>=0.0.1a32",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fdf"]

[project.scripts]
fdf = "fdf.cli.main:main"

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["fdf"]
